All Tests Fixed and Working
