+++ Include/Misra/Std/Container/List/Private.h
@@ -23,5 +23,6 @@ GenericListNode *node_at_list(GenericList *list, u64 item_size, u64 idx);
 void            *item_ptr_at_list(GenericList *list, u64 item_size, u64 idx);
 void             validate_list(const GenericList *list);
 GenericListNode *get_node_relative_to_list_node(GenericListNode *node, i64 ridx);
+GenericListNode *get_node_random_access(GenericList *list, GenericListNode *node, u64 nidx, i64 ridx);
 
 #endif // MISRA_STD_CONTAINER_LIST_PRIVATE_H
