+++ meson.build
@@ -22,6 +22,7 @@ misra_std_sources = files(
     'Source/Misra/Std/Memory.c',
     'Source/Misra/Std/Container/Vec.c',
     'Source/Misra/Std/Container/Str.c',
+    'Source/Misra/Std/Container/List.c',
     'Source/Misra/Std/Container/BitVec.c',
     'Source/Misra/Std/Utility/Iter.c',
     'Source/Misra/Parsers/JSON.c'
